requirements.txt
python-telegram-bot>=20.0,<21.0
openai>=1.0.0
python-dotenv>=1.0.0
